+++ raphodo/rpdsql.py
@@ -355,8 +355,12 @@ class ThumbnailRowsSQL:
         row = self.conn.execute('SELECT uid FROM files WHERE marked=1 LIMIT 1').fetchone()
         return row is not None
 
-    def any_files_to_download(self) -> bool:
-        row = self.conn.execute('SELECT uid FROM files WHERE downloaded=0 LIMIT 1').fetchone()
+    def any_files_to_download(self, scan_id: Optional[int]=None) -> bool:
+        if scan_id is not None:
+            row = self.conn.execute('SELECT uid FROM files WHERE downloaded=0 AND scan_id=? '
+                                    'LIMIT 1', (scan_id,)).fetchone()
+        else:
+            row = self.conn.execute('SELECT uid FROM files WHERE downloaded=0 LIMIT 1').fetchone()
         return row is not None
 
     def any_files_with_extensions(self, scan_id: int, extensions: List[str]) -> bool: